+++ module/bar3/tokenHud.js
@@ -1,5 +1,9 @@
 /** Large portions have been inspired by https://gitlab.com/woodentavern/foundryvtt-bar-brawl */
 
+import { getBarExtendedAttribute } from "./extendTokenConfig.js"
+
+// TokenHUD does not appear to be extendable, so hook on renderTokenHUD
+
 /**
  * Modifies the given HTML to render additional resource input fields.
  * @param {TokenHUD} tokenHud The HUD object.
@@ -7,6 +11,11 @@
  * @param {Object} data The data of the token HUD.
  */
 export const HeroSystem6eTokenHud = async function (tokenHud, html, data) {
+
+    // Make sure were alphaTesting
+    // if (!game.settings.get(game.system.id, 'alphaTesting')) {
+    //     return html;
+    // }
 
     // TokenHud includes most of the clickable controls/inputs
     // it does not include the bar's themselves.
@@ -18,28 +27,17 @@ export const HeroSystem6eTokenHud = async function (tokenHud, html, data) {
     middleColumn.find("div.attribute").remove();
     //html.find(".col.middle .bar1").remove()
 
-    // Do not now how data is created, but it has bar1Data and bar2Data.
-    // We will create bar3Data here for now.
-    const token = tokenHud.object
-    const actor = token?.actor
-    if (actor) {
-        data.bar3Data = {
-            attribute: 'characteristics.end',
-            value: actor.system.characteristics.end.value,
-            max: actor.system.characteristics.end.value,
-            type: "heroBar",
-            name: "heroBar3",
-            editable: true
-        }
-        //await token.document.update({bar3: data.bar3Data.attribute})
-    } else {
-        return /* forget about all the custom stuff */
-    }
+    let actor = tokenHud.actor || tokenHud.object.document.getActor()
+    data.actor = actor
 
+    // Define bar3data
+    const bar3 = getBarExtendedAttribute.bind(data)("bar3") //tokenHud.object.document.getBarAttribute("bar3")
+    data.bar3Data = bar3;
+    data.displayBar3 = bar3 && (bar3.type !== "none");
 
     // Add top bar3
     let bars = []
-    bars.push(data.bar3Data)
+    bars.push(data.bar3Data || {})
     bars[0].name = "bar3"
     bars[0].maxcolor = "blue"
     for (let i = 0; i < bars.length; i++) {
@@ -49,10 +47,10 @@ export const HeroSystem6eTokenHud = async function (tokenHud, html, data) {
 
     // Add bottom bar1 & bar2
     bars = []
-    bars.push(data.bar1Data)
+    bars.push(data.bar1Data || {})
     bars[0].name = "bar1"
     bars[0].maxcolor = "#CC5500"
-    bars.push(data.bar2Data)
+    bars.push(data.bar2Data || {})
     bars[1].name = "bar2"
     bars[1].maxcolor = "green"
     for (let i = 0; i < bars.length; i++) {
@@ -60,11 +58,43 @@ export const HeroSystem6eTokenHud = async function (tokenHud, html, data) {
     }
     middleColumn.append(await renderBarInput(bars))
 
-    // Add input events
+    // Add input events (__onAttributeUpdate is broken for bar3)
     html.find(".attribute.hero-attribute input")
         .click(tokenHud._onAttributeClick)
         .keydown(tokenHud._onAttributeKeydown.bind(tokenHud))
-        .focusout(tokenHud._onAttributeUpdate.bind(tokenHud));
+        .focusout(_onAttributeUpdate.bind(tokenHud));
+
+    async function _onAttributeUpdate(event) {
+        event.preventDefault();
+        if (!this.object) return;
+
+        // Acquire string input
+        const input = event.currentTarget;
+        let strVal = input.value.trim();
+        let isDelta = strVal.startsWith("+") || strVal.startsWith("-");
+        if (strVal.startsWith("=")) strVal = strVal.slice(1);
+        let value = Number(strVal);
+
+        // For attribute bar values, update the associated Actor
+        const bar = input.dataset.bar;
+        const actor = this.object?.actor;
+        if (bar && actor) {
+            const attr = getBarExtendedAttribute.bind(this.object.document)("bar3"); //this.object.document.getBarAttribute(bar);
+            if (isDelta || (attr.attribute !== value)) {
+                actor.modifyTokenAttribute(attr.attribute, value, isDelta, attr.type === "bar");
+            }
+        }
+
+        // Otherwise update the Token directly
+        else {
+            const current = foundry.utils.getProperty(this.object.document, input.name);
+            this.object.document.update({ [input.name]: isDelta ? current + value : value });
+        }
+
+        // Clear the HUD
+        //if ( !this.#hoverControlIcon ) this.clear();
+        this.clear();
+    }
 
 }
 
@@ -76,7 +106,7 @@ export const HeroSystem6eTokenHud = async function (tokenHud, html, data) {
  */
 function renderBarInput(bars) {
     const css = bars.length > 1 ? "compact" : ""
-    return renderTemplate("systems/hero6efoundryvttv2/templates//resource-hud.hbs", { bars, css });
+    return renderTemplate("systems/hero6efoundryvttv2/module/bar3/resource-hud.hbs", { bars, css });
 }